requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
aotools==1.0.6
astropy==5.0.4
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.21.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
kiwisolver==1.4.2
llvmlite==0.38.0
Markdown==3.3.6
matplotlib==3.5.2
multidict==6.0.2
names==0.3.0
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.1.0
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyerfa==2.0.0.1
pyFFTW==0.13.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytorch-lightning==1.6.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soapy==0.13.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==0.8.1
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
wandb==0.12.15
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
